requirements.txt
streamlit
langchain
langchain_community
langchain_core
faiss-cpu
ollama
python-dotenv